requirements.txt
fastapi
uvicorn
httpx
python-dotenv
slack-bolt
aiohttp
requests
pandas
databricks-sdk